# packages/api-contract/openapi/clumsies.admin.v1.yaml
openapi: 3.1.0
info:
  title: clumsies Admin API
  version: 1.0.0
  summary: Admin Server API for native organization administration.
servers:
  - url: https://server.clumsies.local
security:
  - bearerAuth: []
paths:
  /api/v1/setup:
    get:
      operationId: getSetup
      summary: Read the one-time Server setup state.
      security: []
      responses:
        "200":
          description: Current setup state and any active setup session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/setup/sessions:
    post:
      operationId: createSetupSession
      summary: Exchange the deployment Setup Code for a short setup session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSetupSessionRequest"
      responses:
        "201":
          description: Setup session created. The session credential is returned as an HttpOnly cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSetupSessionResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/setup/configuration:
    put:
      operationId: replaceSetupConfiguration
      summary: Replace the organization and default project configuration for this setup session.
      security:
        - setupSession: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceSetupConfigurationRequest"
      responses:
        "200":
          description: Saved setup configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupConfiguration"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/setup/oidc-authorizations:
    post:
      operationId: createSetupOidcAuthorization
      summary: Create the OIDC authorization used to bind the first Owner.
      security:
        - setupSession: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupOidcAuthorizationRequest"
      responses:
        "201":
          description: OIDC authorization created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupOidcAuthorization"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/identity-provider:
    get:
      operationId: getAdminIdentityProvider
      summary: Read the safe enterprise identity provider configuration.
      responses:
        "200":
          description: OIDC Provider state without credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcProviderStatus"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/org:
    get:
      operationId: getAdminOrg
      summary: Read organization settings.
      responses:
        "200":
          description: Organization settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrg"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateAdminOrg
      summary: Update organization settings.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminOrgRequest"
      responses:
        "200":
          description: Updated organization settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrg"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/members:
    get:
      operationId: listAdminMembers
      summary: List organization members.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Member page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createAdminMember
      summary: Create an organization member admission record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemberRequest"
      responses:
        "201":
          description: Created member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    patch:
      operationId: updateAdminMember
      summary: Update organization member role or status.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Updated member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAdminMember
      summary: Remove or disable an organization member.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Delete result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/projects:
    get:
      operationId: listAdminProjects
      summary: List project metadata for administration.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Project metadata page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProjectListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createAdminProject
      summary: Create a project and grant the current administrator project admin access.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: getAdminProject
      summary: Read project metadata for administration.
      responses:
        "200":
          description: Project metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateAdminProject
      summary: Update project metadata.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAdminProject
      summary: Delete a project and its governed content.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Delete result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/projects/{project_id}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listAdminProjectMembers
      summary: List project members.
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/ProjectRole"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Project member page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createAdminProjectMember
      summary: Add an organization member to a project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectMemberRequest"
      responses:
        "201":
          description: Created project membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/projects/{project_id}/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/UserId"
    patch:
      operationId: updateAdminProjectMember
      summary: Update a project member role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectMemberRequest"
      responses:
        "200":
          description: Updated project membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMember"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteAdminProjectMember
      summary: Remove a member from a project.
      responses:
        "200":
          description: Delete result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    setupSession:
      type: apiKey
      in: cookie
      name: __Host-clumsies_setup_session
      description: >-
        HttpOnly one-time setup session cookie. Local plain-HTTP development
        uses the unprefixed clumsies_setup_session name because __Host-
        cookies require HTTPS.
  parameters:
...
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SetupStatus:
      type: object
      required: [state, setup_code_configured, oidc_configured, session]
      properties:
        state:
          $ref: "#/components/schemas/InstallationState"
        setup_code_configured:
          type: boolean
        oidc_configured:
          type: boolean
        session:
          oneOf:
            - $ref: "#/components/schemas/SetupSessionStatus"
            - type: "null"
    CreateSetupSessionRequest:
      type: object
      required: [setup_code]
      properties:
        setup_code:
          type: string
          minLength: 1
    CreateSetupSessionResponse:
      type: object
      required: [expires_at, csrf_token]
      properties:
        expires_at:
          type: string
          format: date-time
        csrf_token:
          type: string
    ReplaceSetupConfigurationRequest:
      allOf:
        - $ref: "#/components/schemas/SetupConfiguration"
...
    SetupOidcAuthorization:
      type: object
      required: [authorization_url]
      properties:
        authorization_url:
          type: string
          format: uri
...